# contracts/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AccessForge API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workspaces/{workspace_id}/runners": {
      "get": {
        "description": "The runner inventory, with the evidence behind each status.\n\n`preflightPassedAt` is the field that matters, and it is separate from `status` on purpose. A\nrunner process being reachable is not proof that a screen reader is running on it, that the\nreader is the one enrolled, or that anything was ever read back from a real desktop. INV-02\nturns on that distinction: a missing reader capability must never resolve to a pass, and a UI\nthat inferred readiness from \"the row exists\" would be the first place that inference is made.\n\nThe reset counter is included because a desktop that keeps needing resets is a desktop nobody\nshould be scheduling work onto, and that pattern is invisible from a single status word.",
        "operationId": "list_runners_v1_workspaces__workspace_id__runners_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Runners V1 Workspaces  Workspace Id  Runners Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runners",
        "tags": [
          "runners"
        ]
      },
      "post": {
        "operationId": "enroll_runner_v1_workspaces__workspace_id__runners_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enroll Runner V1 Workspaces  Workspace Id  Runners Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enroll Runner",
        "tags": [
          "runners"
        ]
      }
    },
    "/v1/workspaces/{workspace_id}/runners/enrollment-tokens": {
      "post": {
        "description": "Issue a short-lived single-use enrollment credential.\n\nThe token is returned exactly once, here. There is no route that reads one back: an endpoint\nthat could reveal an enrollment token would be a way to take over a tenant's desktop, and\n\"only administrators can call it\" is a weaker property than \"it does not exist\".",
        "operationId": "issue_enrollment_token_v1_workspaces__workspace_id__runners_enrollment_tokens_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Issue Enrollment Token V1 Workspaces  Workspace Id  Runners Enrollment Tokens Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue Enrollment Token",
        "tags": [
          "runners"
        ]
      }
    },
    "/v1/workspaces/{workspace_id}/runners/{runner_id}": {
      "get": {
        "description": "Everything an operator needs before deciding whether a reset is safe.\n\nIncludes the unresolved actions and the local-impact warning, because \"is anything still in\nflight\" is what separates a reset that is an inconvenience from one that interrupts work.",
        "operationId": "inspect_runner_v1_workspaces__workspace_id__runners__runner_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "runner_id",
            "required": true,
            "schema": {
              "title": "Runner Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Inspect Runner V1 Workspaces  Workspace Id  Runners  Runner Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect Runner",
        "tags": [
          "runners"
        ]
      }
    },
    "/v1/workspaces/{workspace_id}/runners/{runner_id}/resets": {
      "post": {
        "description": "Attempt the trusted reset that is the only route out of quarantine.\n\n`succeeded` comes from the reset procedure, never from the runner. A failed reset leaves the\ndesktop quarantined: failing to prove the old actor cannot act is not proof that it cannot.",
        "operationId": "reset_runner_v1_workspaces__workspace_id__runners__runner_id__resets_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "runner_id",
            "required": true,
            "schema": {
              "title": "Runner Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reset Runner V1 Workspaces  Workspace Id  Runners  Runner Id  Resets Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Runner",
        "tags": [
          "runners"
        ]
      }
    },
...
  }
}